+++ leggedsnake/physicsengine.py
@@ -14,13 +14,10 @@ Created on Sat May 25 2019 14:56:01.
 """
 import numpy as np
 import pymunk as pm
-import pymunk.matplotlib_util
-import matplotlib.pyplot as plt
-import matplotlib.animation as anim
 from pylinkage.geometry import norm, cyl_to_cart, bounding_box
 from pylinkage.linkage import Static, Crank, Fixed, Pivot
 
-from . import dynamiclinkage as dlink
+from . import dynamiclinkage
 
 # Simulation parameters
 params = {
@@ -58,16 +55,9 @@ params = {
     },
     # Study hypothesis
     "simul": {
-        # Time between two physics compuation
+        # Time between two physics computations
         "physics_period": 0.02,
-    },
-    # Display parameters
-    "camera": {
-        # Do you want to follow a system of view whole scene?
-        "dynamic_camera": True,
-        # Required frames per second
-        "fps": 20,
-    },
+    }
 }
 
 
@@ -79,8 +69,9 @@ def set_space_constraints(space):
     space.iterations = int(10 * np.exp(len_c / 60))
     for constraint in constraints:
         if not isinstance(constraint, pm.SimpleMotor) and False:
-            constraint.max_force = params["physics"]["max_force"] * (np.exp(
-                - len_c / 25) / 2 + .5)
+            constraint.max_force = params["physics"]["max_force"] * (
+                    np.exp(- len_c / 25) / 2 + .5
+            )
         constraint.error_bias = (1 - .1 * np.exp(-len_c / 60)) ** 60
 
 
@@ -117,22 +108,40 @@ class World:
         # The road which will be built
         self.road = [(-15, road_y), (15, road_y)]
         # First road parts
-        seg = pm.Segment(self.space.static_body, self.road[0], self.road[-1],
-                         .1)
+        seg = pm.Segment(
+            self.space.static_body, self.road[0], self.road[-1],
+            .1
+        )
         seg.friction = params["ground"]["friction"]
         self.space.add(seg)
         self.linkages = []
 
-    def add_linkage(self, linkage):
-        """Add a DynamicLinkage to the simulation."""
-        if isinstance(linkage, dlink.DynamicLinkage):
-            dynamiclinkage = linkage
+    def add_linkage(self, linkage, load=0):
+        """
+        Add a DynamicLinkage to the simulation.
+
+        Parameters
+        ----------
+        linkage : pylinkage.Linkage or leggedsnake.DynamicLinkage
+            Linkage to add.
+            We use the linkage's space if it is a DynamicLinkage.
+        load : float, optional
+            Load to add the center of the linkage.
+            Has no effect when using a DynamicLinkage.
+            The default is 0.
+        Returns
+        -------
+
+        """
+        if isinstance(linkage, dynamiclinkage.DynamicLinkage):
+            dynamic_linkage = linkage
         else:
-            dynamiclinkage = dlink.convert_to_dynamic_linkage(
-                linkage, self.space)
-        for crank in dynamiclinkage._cranks:
-            crank.actuator.max_force = 0
-        self.linkages.append(dynamiclinkage)
+            dynamic_linkage = dynamiclinkage.convert_to_dynamic_linkage(
+                linkage, self.space, load=load
+            )
+        for cur_crank in dynamic_linkage._cranks:
+            cur_crank.actuator.max_force = 0
+        self.linkages.append(dynamic_linkage)
         for s in self.space.shapes:
             s.friction = params["ground"]["friction"]
         # set_space_constraints(self.space)
@@ -142,11 +151,13 @@ class World:
         linkage_crank = next(j for j in linkage.joints if isinstance(j, Crank))
         if (
                 linkage_crank.actuator.max_force == 0
-                and norm(linkage.body.velocity) < .1):
+                and norm(linkage.body.velocity) < .1
+        ):
             linkage_crank.actuator.max_force = params["linkage"]["torque"]
             linkage.height = linkage.body.position.y
-            linkage.mechanical_energy = (.5 * linkage.mass
-                                         * norm(linkage.body.velocity) ** 2)
+            linkage.mechanical_energy = (
+                    .5 * linkage.mass * norm(linkage.body.velocity) ** 2
+            )
 
         # Energy from the motor in this step
         energy = power * params["simul"]["physics_period"]
@@ -154,13 +165,12 @@ class World:
             v = norm(linkage.body.velocity)
             g = norm(params["physics"]["gravity"])
             m = linkage.mass
-            new_mechanical_energy = (m
-                                     * (.5 * v ** 2
-                                        + g * (linkage.body.position.y
-                                               - linkage.height))
-                                     )
-            efficiency = (new_mechanical_energy
-                          - linkage.mechanical_energy) / energy
+            new_mechanical_energy = m * (
+                .5 * v ** 2 + g * (linkage.body.position.y - linkage.height)
+            )
+            efficiency = (
+                new_mechanical_energy - linkage.mechanical_energy
+            ) / energy
             linkage.mechanical_energy = new_mechanical_energy
             return energy, efficiency
         return 0, 0
@@ -198,12 +208,16 @@ class World:
         energies = [0] * len(self.linkages)
         efficiencies = [0] * len(self.linkages)
         for i, linkage, power in zip(
-                range(len(self.linkages)), self.linkages, powers):
+                range(len(self.linkages)), self.linkages, powers
+        ):
             recalc_linkage(linkage)
-            energies[i], efficiencies[i] = self.__update_linkage__(linkage,
-                                                                   power[0])
-            bounds = (min(bounds[0], *(i.x for i in linkage.joints)),
-                      max(bounds[1], *(i.x for i in linkage.joints)))
+            energies[i], efficiencies[i] = self.__update_linkage__(
+                linkage, power[0]
+            )
+            bounds = (
+                min(bounds[0], *(i.x for i in linkage.joints)),
+                max(bounds[1], *(i.x for i in linkage.joints))
+            )
         while self.road[-1][0] < bounds[1] + 10:
             self.build_road(True)
         while self.road[0][0] > bounds[0] - 10:
@@ -212,15 +226,18 @@ class World:
         # Without animation, we return 100 times motor yield
         # with a duration step
         for linkage, energy, efficiency in zip(
-                self.linkages, energies, efficiencies):
+                self.linkages, energies, efficiencies
+        ):
             return efficiency, energy * dt
 
     def __build_road_step__(self, ground, index):
         """Add a step (two points)."""
         high = np.random.rand() * ground["max_step"]
         a = self.road[index][0], self.road[index][1] + high
-        b = (self.road[index][0] + ground["section_len"] * (1 - index),
-             self.road[index][1] + high)
+        b = (
+            self.road[index][0] + ground["section_len"] * (1 - index),
+            self.road[index][1] + high
+        )
 
         s = pm.Segment(self.space.static_body, a, b, .1)
         s.friction = ground["friction"]
@@ -235,9 +252,10 @@ class World:
     def __build_road_segment__(self, ground, index):
         """Add a segment (one point)."""
         # Add noise for more chaotic terrain."""
-        angle = np.random.normal(ground["slope"] / 2,
-                                 ground["noise"] * ground["slope"] / 2)
-        # Adding a point to the left is increasing angle by pi/2
+        angle = np.random.normal(
+            ground["slope"] / 2, ground["noise"] * ground["slope"] / 2
+        )
+        # Adding a point to the left increases angle by pi/2
         if not index:
             angle = np.pi - angle
         a = pm.Vec2d(*cyl_to_cart(ground["section_len"], angle,
@@ -263,145 +281,6 @@ class World:
             self.__build_road_segment__(ground, -positive)
 
 
-class VisualWorld(World):
-    """Same as parent class World, but with matplotlib objects."""
-
-    def __init__(self, space=None, road_y=-5):
-        """
-        Instantiate the world and objects to be displayed.
-
-        Parameters
-        ----------
-        space : pymunk.space.Space, optional
-            Space of simulation. The default is None.
-        road_y : float, optional
-            The ordinate of the ground. Useful when likages have long legs.
-            The default is -5.
-        """
-        super().__init__(space=space, road_y=road_y)
-        self.fig, self.ax = plt.subplots()
-        self.ax.set_aspect('equal')
-        self.linkage_im = []
-        # Same for the road
-        self.road_im = self.ax.plot([], [], 'k-', animated=False)
-
-    def add_linkage(self, linkage):
-        """
-        Add a linkage to the world, and create the appropriate Artist objects.
-
-        Parameters
-        ----------
-        linkage : pylinkage.linkage.Linkage
-            The linkage you want to add, can be a Walker.
-
-        Returns
-        -------
-        None.
-
-        """
-        super().add_linkage(linkage)
-        # Objects that will allow display
-        linkage_im = []
-        ax = self.ax
-        for j in self.linkages[-1].joints:
-            if (
-                    isinstance(j, Static)
-                    and hasattr(j, 'joint0')
-                    and j.joint0 is not None):
-                linkage_im.append(ax.plot([], [], 'k-', animated=False)[0])
-                if hasattr(j, 'joint1') and j.joint1 is not None:
-                    linkage_im.append(ax.plot([], [], 'k-', animated=False)[0])
-            elif isinstance(j, Crank):
-                linkage_im.append(ax.plot([], [], 'g-', animated=False)[0])
-            elif isinstance(j, Fixed):
-                linkage_im.append(ax.plot([], [], 'r-', animated=False)[0])
-                linkage_im.append(ax.plot([], [], 'r-', animated=False)[0])
-            elif isinstance(j, Pivot):
-                linkage_im.append(ax.plot([], [], 'b-', animated=False)[0])
-                linkage_im.append(ax.plot([], [], 'b-', animated=False)[0])
-        self.linkage_im.extend(linkage_im)
-
-    def draw_linkage(self, joints):
-        """Draw the linkage at his current state."""
-        a = 0
-        for j in joints:
-            if hasattr(j, 'joint0') and j.joint0 is not None:
-                self.linkage_im[a].set_data(
-                    [j.x, j.joint0.x], [j.y, j.joint0.y]
-                )
-                a += 1
-            if hasattr(j, 'joint1') and j.joint1 is not None:
-                self.linkage_im[a].set_data(
-                    [j.x, j.joint1.x], [j.y, j.joint1.y]
-                )
-                a += 1
-        return self.linkage_im
-    
-    def reload_visuals(self):
-        """Reload the visual components only."""
-        center = self.linkages[0].joints[0].coord()
-        self.fig.suptitle(f"Position: {tuple(map(int, center))}")
-
-        self.road_im[0].set_data(
-            [i[0] for i in self.road],
-            [i[1] for i in self.road]
-        )
-        ax = self.ax
-        if params["camera"]["dynamic_camera"]:
-            ax.set_xlim(center[0] - 10, center[0] + 10)
-            ax.set_ylim(center[1] - 10, center[1] + 10)
-        else:
-            ax.set_ylim(
-                min([0] + [min(i.y for i in linkage.joints) for linkage in self.linkages]) - 5,
-                max([0] + [max(i.y for i in linkage.joints) for linkage in self.linkages]) + 5
-            )
-
-        # Return modified objects for animation optimization
-        visual_objects = []
-        joints = []
-        for linkage in self.linkages:
-            joints.extend(linkage.joints)
-        visual_objects += self.draw_linkage(joints)
-        visual_objects += self.road_im
-        return visual_objects
-
-    def visual_update(self, time=None):
-        """
-        Update simulation and draw it.
-        
-        Parameters
-        ----------
-        time : list | float | None
-            When a list, delta-time for physics and display (respectively) 
-            Using a float, only delta-time for physics, fps is set with params["camera"]["fps"]
-            Setting to None set physics dt to params["simul"]["physics_period"] and fps to params["camera"]["fps"]
-        """
-        if time is None:
-            dt = params["simul"]["physics_period"]
-            fps = params["camera"]["fps"]
-        elif isinstance(time, int) or isinstance(time, float):
-            dt = time
-            fps = params["camera"]["fps"]
-        else:
-            dt, fps = time
-        div = 1 // (dt * fps)
-        if div >= 1:
-            update_ret = [0, 0]
-            for _ in range(int(div)):
-                for i, step_update in enumerate(self.update(dt)):
-                    update_ret[i] += step_update
-            for i, step_update in enumerate(self.update(1 / fps - dt * div)):
-                update_ret[i] += step_update
-        else:
-            print(
-                f"Warning: Physics is computed every {dt}s ({1 / dt} times/s)",
-                f"but display is {fps} times/s."
-            )
-            update_ret = self.update(dt)
-        self.reload_visuals()
-        return update_ret
-
-
 def recalc_linkage(linkage):
     """Assign a good position to all joints."""
     for j in linkage.joints:
@@ -420,126 +299,9 @@ def linkage_bb(linkage):
         The linkage from which to get the bounding box.
     """
     data = [i.coord() for i in linkage.joints]
-    if isinstance(linkage, dlink.DynamicLinkage):
+    if isinstance(linkage, dynamiclinkage.DynamicLinkage):
         data.extend(tuple(i.position) for i in linkage.rigidbodies)
     return bounding_box(data)
-
-
-def im_debug(world, linkage):
-    """Use pymunk debugging for visual debugging."""
-    bbox = linkage_bb(linkage)
-    world.ax.clear()
-    world.ax.set_xlim(int(bbox[3]) - 5, int(bbox[1]) + 5)
-    world.ax.set_ylim(int(bbox[2]) - 5, int(bbox[0]) + 5)
-    world.ax.scatter(
-        [i.x for i in linkage.joints],
-        [i.y for i in linkage.joints],
-        c='r'
-    )
-    for j in linkage.joints:
-        for shape in j._a.shapes:
-            begin = j._a.local_to_world(shape.a)
-            end = j._a.local_to_world(shape.b)
-            world.ax.plot([begin[0], end[0]], [begin[1], end[1]])
-    options = pymunk.matplotlib_util.DrawOptions(world.ax)
-    options.constraint_color = (.1, .1, .1, .0)
-    world.space.debug_draw(options)
-
-
-def video_debug(linkage):
-    """Launch the simulation frame by frame, useful for debug."""
-    road_y = linkage_bb(linkage)[0] - 1
-    if isinstance(linkage, dlink.DynamicLinkage):
-        world = VisualWorld(linkage.space, road_y=road_y)
-    else:
-        world = VisualWorld(road_y=road_y)
-    world.add_linkage(linkage)
-    dynamic_linkage = world.linkages[-1]
-    for _ in range(1, int(1e3)):
-        dt = params["simul"]["physics_period"]
-        world.space.step(dt)
-        recalc_linkage(dynamic_linkage)
-        im_debug(world, dynamic_linkage)
-        plt.pause(.2)
-
-
-def video(linkage, duration=30, save=False):
-    """
-    Give the rigidbody a dynamic model and launch simulation with video.
-
-    Parameters
-    ----------
-    linkage : Union[pylinkage.linkage.Linkage,
-    leggedsnake.dynamiclinkage.DynamicLinkage]
-        The Linkage you want to simulate.
-    duration : float, optional
-        Duration (in seconds) of the simulation. The default is 40.
-    save : bool, optional
-        If you want to save it as a .mp4 file.
-    """
-    road_y = linkage_bb(linkage)[0] - 1
-    if isinstance(linkage, dlink.DynamicLinkage):
-        world = VisualWorld(linkage.space, road_y=road_y)
-    else:
-        world = VisualWorld(road_y=road_y)
-    world.add_linkage(linkage)
-    # Number of frames for the selected duration
-    n_frames = int(params["camera"]["fps"] * duration)
-
-    animation = anim.FuncAnimation(
-        world.fig, world.visual_update,
-        frames=[None] * (n_frames - 1),
-        interval=int(1000 / params["camera"]["fps"]),
-        repeat=False, blit=False
-    )
-    if save:
-        writer = anim.FFMpegWriter(fps=params["camera"]["fps"], bitrate=2500)
-        animation.save(f"Dynamic {linkage.name}.mp4", writer=writer)
-    else:
-        plt.show()
-        if animation:
-            pass
-
-
-def all_linkages_video(linkages, duration=30, save=False):
-    """
-    Give the rigidbody a dynamic model and launch simulation with video.
-
-    Parameters
-    ----------
-    linkages : Union[
-        list of pylinkage.linkage.Linkage,
-        list of leggedsnake.dynamiclinkage.DynamicLinkage
-    ]
-        The Linkage you want to simulate.
-    duration : float, optional
-        Duration (in seconds) of the simulation. The default is 40.
-    save : bool, optional
-        If you want to save it as a .mp4 file.
-    """
-    road_y = linkage_bb(linkages[0])[0] - 1
-    if isinstance(linkages[0], dlink.DynamicLinkage):
-        world = VisualWorld(linkages[0].space, road_y=road_y)
-    else:
-        world = VisualWorld(road_y=road_y)
-    for linkage in linkages:
-        world.add_linkage(linkage)
-    # Number of frames for the selected duration
-    n_frames = int(params["camera"]["fps"] * duration)
-
-    animation = anim.FuncAnimation(
-        world.fig, world.visual_update,
-        frames=[None] * (n_frames - 1),
-        interval=int(1000 / params["camera"]["fps"]),
-        repeat=False, blit=False
-    )
-    if save:
-        writer = anim.FFMpegWriter(fps=params["camera"]["fps"], bitrate=2500)
-        animation.save(f"Dynamic {linkage[0].name}.mp4", writer=writer)
-    else:
-        plt.show()
-        if animation:
-            pass
 
 
 if __name__ == "__main__":
@@ -548,10 +310,9 @@ if __name__ == "__main__":
     follower = Pivot(0, 2, joint0=base, joint1=crank, distance0=2,
                      distance1=1)
     frame = Fixed(joint0=crank, joint1=follower, distance=1, angle=-np.pi/2)
-    demo_linkage = dlink.DynamicLinkage(
+    demo_linkage = dynamiclinkage.DynamicLinkage(
         name='Some tricky linkage',
         joints=(base, crank, follower, frame),
         space=pm.Space()
     )
     demo_linkage.space.gravity = params["physics"]["gravity"]
-    video_debug(demo_linkage)